packages/backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
python-multipart==0.0.20
slowapi==0.1.9
psycopg2-binary==2.9.11
bcrypt==5.0.0
pyjwt==2.10.1
requests==2.32.5
google-cloud-texttospeech==2.33.0
alembic==1.14.0
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
